Preparing version 7 for wide release
