+++ MedoDelirioBrasilia.xcodeproj/xcshareddata/xcschemes/MedoDelirioBrasilia.xcscheme
@@ -64,7 +64,7 @@
       <CommandLineArguments>
          <CommandLineArgument
             argument = "-UNDER_DEVELOPMENT"
-            isEnabled = "YES">
+            isEnabled = "NO">
          </CommandLineArgument>
          <CommandLineArgument
             argument = "-FORCE_SHOW_HELP_THE_APP"
