+++ MedoDelirioBrasilia/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "beta_icon_2.png",
+      "filename" : "1024.png",
       "idiom" : "universal",
       "platform" : "ios",
       "size" : "1024x1024"
